fix: divide App to LogOnApp, LogOut App
