+++ fashion-today_front/src/app/LogOutApp.js
@@ -1,0 +1,9 @@
+import React from "react";
+
+function LogOutApp() {
+  return (
+    <h1>LogIn</h1>    
+  )
+}
+
+export default LogOutApp;
